timezones in setup_sample_courses.py
