+++ .setup/bin/setup_sample_courses.py
@@ -16,7 +16,7 @@ just the ones used for testing.
 from __future__ import print_function, division
 import argparse
 from collections import OrderedDict
-from datetime import datetime, timedelta
+from datetime import datetime
 import glob
 import grp
 import hashlib
@@ -24,7 +24,6 @@ import json
 import os
 import pwd
 import random
-import re
 import shutil
 import subprocess
 import uuid
@@ -62,7 +61,7 @@ DB_PASS = "hsdbu"
 
 DB_ONLY = False
 
-NOW = datetime.now()
+NOW = submitty_utils.get_current_time()
 
 
 def main():
@@ -142,7 +141,7 @@ def main():
                               user_preferred_firstname=user.preferred_firstname,
                               user_lastname=user.lastname,
                               user_email=user.email,
-                              last_updated=NOW.strftime("%Y-%m-%d %H:%M:%S"))
+                              last_updated=NOW.strftime("%Y-%m-%d %H:%M:%S%z"))
 
     for user in extra_students:
         submitty_conn.execute(user_table.insert(),
@@ -152,7 +151,7 @@ def main():
                               user_preferred_firstname=user.preferred_firstname,
                               user_lastname=user.lastname,
                               user_email=user.email,
-                              last_updated=NOW.strftime("%Y-%m-%d %H:%M:%S"))
+                              last_updated=NOW.strftime("%Y-%m-%d %H:%M:%S%z"))
     submitty_conn.close()
 
     today = datetime.today()
@@ -377,57 +376,10 @@ def get_current_semester():
     return semester
 
 
-def parse_datetime(date_string):
-    """
-    Given a string that should either represent an absolute date or an arbitrary date, parse this
-    into a datetime object that is then used. Absolute dates should be in the format of
-    YYYY-MM-DD HH:MM:SS while arbitrary dates are of the format "+/-# day(s) [at HH:MM:SS]" where
-    the last part is optional. If the time is omitted, then it uses midnight of whatever day was
-    specified.
-
-    Examples of allowed strings:
-    2016-10-14
-    2016-10-13 22:11:32
-    -1 day
-    +2 days at 00:01:01
-
-    :param date_string:
-    :type date_string: str
-    :return:
-    :rtype: datetime
-    """
-    if isinstance(date_string, datetime):
-        return date_string
-    try:
-        return datetime.strptime(date_string, "%Y-%m-%d %H:%M:%S%z")
-    except ValueError:
-        pass
-
-    try:
-        return datetime.strptime(date_string, "%Y-%m-%d %H:%M:%S%z").replace(hour=23, minute=59, second=59)
-    except ValueError:
-        pass
-
-    m = re.search('([+|\-][0-9]+) (days|day) at [0-2][0-9]:[0-5][0-9]:[0-5][0-9]', date_string)
-    if m is not None:
-        hour = int(m.group(2))
-        minu = int(m.group(3))
-        sec = int(m.group(4))
-        days = int(m.group(1))
-        return NOW.replace(hour=hour, minute=minu, second=sec) + timedelta(days=days)
-
-    m = re.search('([+|\-][0-9]+) (days|day)', date_string)
-    if m is not None:
-        days = int(m.group(1))
-        return NOW.replace(hour=23, minute=59, second=59) + timedelta(days=days)
-
-    raise ValueError("Invalid string for date parsing: " + str(date_string))
-
-
 def datetime_str(datetime_obj):
     if not isinstance(datetime_obj, datetime):
         return datetime_obj
-    return datetime_obj.strftime("%Y-%m-%d %H:%M:%S%z")
+    return datetime_obj.strftime('%Y-%m-%d %H:%M:%S%z')
 
 
 def parse_args():
@@ -856,7 +808,7 @@ class Course(object):
                                     score = random.randint(0, component.max_value * 2) / 2
                                 else:
                                     score = random.randint(component.max_value * 2, 0) / 2
-                                grade_time = gradeable.grade_start_date.strftime("%Y-%m-%d %H:%M:%S")
+                                grade_time = gradeable.grade_start_date.strftime("%Y-%m-%d %H:%M:%S%z")
                                 conn.execute(gradeable_component_data.insert(), gc_id=component.key, gd_id=gd_id,
                                              gcd_score=score, gcd_component_comment="lorem ipsum",
                                              gcd_grader_id=self.instructor.id, gcd_grade_time=grade_time, gcd_graded_version=1)
@@ -1127,12 +1079,12 @@ class Gradeable(object):
         if 'grading_rotating' in gradeable:
             self.grading_rotating = gradeable['grading_rotating']
 
-        self.ta_view_date = parse_datetime(gradeable['g_ta_view_start_date'])
-        self.grade_start_date = parse_datetime(gradeable['g_grade_start_date'])
-        self.grade_released_date = parse_datetime(gradeable['g_grade_released_date'])
+        self.ta_view_date = submitty_utils.parse_datetime(gradeable['g_ta_view_start_date'])
+        self.grade_start_date = submitty_utils.parse_datetime(gradeable['g_grade_start_date'])
+        self.grade_released_date = submitty_utils.parse_datetime(gradeable['g_grade_released_date'])
         if self.type == 0:
-            self.submission_open_date = parse_datetime(gradeable['eg_submission_open_date'])
-            self.submission_due_date = parse_datetime(gradeable['eg_submission_due_date'])
+            self.submission_open_date = submitty_utils.parse_datetime(gradeable['eg_submission_open_date'])
+            self.submission_due_date = submitty_utils.parse_datetime(gradeable['eg_submission_due_date'])
             if 'eg_is_repository' in gradeable:
                 self.is_repository = gradeable['eg_is_repository'] is True
             if self.is_repository and 'eg_subdirectory' in gradeable:
